+++ src/command.ts
@@ -195,7 +195,7 @@ export class Command {
             );
           }
           const duration = new Date().getTime() - start;
-          track(this.name, "success", duration).then(() => process.exit());
+          void track(this.name, "success", duration).then(() => process.exit());
         })
         .catch(async (err) => {
           if (getInheritedOption(options, "json")) {